fixes #29 - lock left side when scrolling over it
When the left sidebar is taller than the page height, it’s impossible
to add a “section break” field (the last field in the list) until you
scroll all the way to the bottom of the page. This solves that, but I
don’t feel like it’s a 100% perfect solution.

If anyone wants to tackle/rethink the general issue of keeping the left
side scrolled properly, please feel free!